+++ table/plain_table_key_coding.cc
@@ -91,7 +91,7 @@ Status PlainTableKeyEncoder::AppendKey(const Slice& key,
 
   Slice key_to_write = key;  // Portion of internal key to write out.
 
-  uint32_t user_key_size = static_cast<uint32_t>(I2ULen(key.size()));
+  uint32_t user_key_size = I2ULen(static_cast<uint32_t>(key.size()));
   if (encoding_type_ == kPlain) {
     if (fixed_user_key_len_ == kPlainTableVariableLength) {
       // Write key length
@@ -150,11 +150,11 @@ Status PlainTableKeyEncoder::AppendKey(const Slice& key,
   // in this buffer to safe one file append call, which takes 1 byte.
   if (parsed_key.sequence == 0 && parsed_key.type == kTypeValue) {
     Status s =
-        file->Append(Slice(key_to_write.data(), I2ULen(key_to_write.size())));
+        file->Append(Slice(key_to_write.data(), I2ULen(static_cast<uint32_t>(key_to_write.size()))));
     if (!s.ok()) {
       return s;
     }
-    *offset += I2ULen(key_to_write.size());
+    *offset += I2ULen(static_cast<uint32_t>(key_to_write.size()));
     meta_bytes_buf[*meta_bytes_buf_size] = PlainTableFactory::kValueTypeSeqId0;
     *meta_bytes_buf_size += 1;
   } else {
